feat!: §28 frames-schema widening — REAL exposure_seconds + nullable gain, in-place DB rebuild (§40 prerequisite) (#670)
* feat!: §28 frames-schema widening — REAL exposure_seconds + nullable gain, with in-place DB rebuild (§40 prerequisite)

The dedicated pass the entry prescribed, unblocking §40 calibration
workflows:

- exposure_seconds is REAL end-to-end: FrameDto/FrameListItemDto/
  DarkLibraryEntryDto double, readers GetDouble, the frames DDL REAL.
  The §28.8 FITS rescan now parses EXPTIME as the double it is — a
  "0.5" header used to FAIL int.TryParse and record 0 (a worse bug
  than the round-up: sub-second orphans recovered as zero-length).
- Gain is nullable end-to-end: the capture path's `Gain ?? -1` sentinel
  is gone; a camera or FITS that reports no gain records NULL. Writes
  use the existing DbValue idiom; reads IsDBNull.
- Old-schema upgrade: SQLite can't ALTER COLUMN, so InitializeAsync
  runs an in-place one-transaction rebuild when the stored frames DDL
  matches the old shape (idempotent DDL-sniff trigger; fresh DBs skip):
  every row copied, legacy -1 gains normalized to NULL in the copy,
  all three frames indexes recreated.
- Stats aggregations already CAST to REAL (untouched); Astrobin CSV
  formats sub-second exposure ("0.####") and blanks unknown gain; the
  client stats CSV formats sub-second durations (inSeconds truncated);
  the openapi Frame schema already said `number`.
- Deliberately out of scope: settings ints (imaging-defaults/autofocus
  exposure) — user-entered whole seconds, not catalog measurements.

Tests: sub-second + null-gain round-trip; the migration test hand-builds
the v0.0.1 schema with a -1 sentinel row, initializes the current build
over it, and asserts row survival, sentinel->NULL, sub-second inserts
now working, all 3 indexes recreated, and idempotent re-initialize.
Server 1453 green, client 1240 green.


* fix: r1 — the REAL capture path no longer rounds sub-second exposures up to 1s

RegisterFrameAsync (the only frame-insert call site for actual camera
captures) still had the legacy Math.Max(1, (int)Math.Round(...)) — the
one line the widening pass existed for. Pass request.ExposureSec through
verbatim; the stale "tracked in PORT_TODO" comment goes with it.

Coverage for the gap the reviewer named (no test drives sub-second
through StartExposureAsync): the §14e OmniSim integration test now
exposes 0.5s and asserts the catalogued frame reads exactly 0.5 with
NULL gain — it runs in the alpaca-sim-integration job against the real
expose→download→FITS→catalog pipeline.


* fix: r2 — AstroBin acquisition CSV honors the widened schema

OpenAstrobinExportAsync still read duration with GetInt32 (a 0.5s
lucky-imaging sub exported duration=0) and gain with no null check
(NULL gain exported a false gain=0). Duration now formats "0.####"
and NULL gain blanks the field, matching the sibling OpenCsvExport.

Test: sub-second + null-gain row asserts the exact CSV line; the
insert helpers widened to double/int? to make it expressible.


---------

+++ OpenAstroAra.Server/Contracts/CalibrationDtos.cs
@@ -78,7 +78,7 @@ public sealed record DarkLibraryStateDto(
 /// <summary>One entry in the dark library.</summary>
 public sealed record DarkLibraryEntryDto(
     Guid Id,
-    int ExposureSeconds,
+    double ExposureSeconds,
     int Gain,
     double TemperatureC,
     int FrameCount,
